requirements.txt
matplotlib

flask

mysql-connector

requests

cloudinary

